[Refactor/feat]PasswordTextField & UserScreen:
PasswordTextfield had minor refactoring so that the password strength and matching messages weren't as far from the text fields, this made it easier to implement on UserScreen and also personally think looks better on forgot password screen too.
The front for profile screen is pretty much done, minor tunings will be applied further on but the code is mostly likely going to be refactored as there is some repetition was 180 lines is near the line limit for a file.